Add GPS functionality

+++ include/packets.h
@@ -27,9 +27,9 @@ typedef struct
     uint8_t flags; // MSB - BOX | BUFFER FULL | NC | NC | FUEL_LEVEL | SERVO_ERROR | CHK | RUN - LSB
     uint8_t soc;
     uint16_t cvt;
-    uint8_t volt;
-    uint16_t latitude;
-    uint16_t longitude;
+    double volt;
+    double latitude;
+    double longitude;
     uint32_t timestamp;
 
 } packet_t;
